+++ packages/shared/src/index.ts
@@ -11,6 +11,7 @@ export * from "./releaseAdapter.js";
 export * from "./graphWalker.js";
 export * from "./approval.js";
 export * from "./approvalGates.js";
+export * from "./approvalPolicy.js";
 export * from "./vegaClient.js";
 export * from "./vegaTransport.js";
 export * from "./ldSdk.js";